requirements.txt
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==5.3.2
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
cv==1.0.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
fpdf==1.7.2
google-auth==2.28.0
GitPython==3.1.43
h11==0.14.0
idna==3.6
Jinja2==3.1.3
kubernetes==29.0.0
MarkupSafe==2.1.5
MouseInfo==0.1.3
numpy==1.26.4
oauthlib==3.2.2
openpyxl==3.1.5
outcome==1.3.0.post0
oracledb==3.2.0
packaging==24.1
pandas==2.2.0
pikepdf==9.5.1
pdfkit==1.0.0
pexpect==4.9.0
pillow==10.2.0
ptyprocess==0.7.0
pyarrow==17.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
PyPDF2==3.0.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytweening==1.2.0
pytz==2024.1
PyYAML==6.0.1
reportlab==4.1.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
rubicon-objc==0.4.9
selenium==4.16.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
termcolor==2.4.0
tomlkit==0.13.3
trcli==1.9.5
trio==0.25.1
trio-websocket==0.11.1
tzdata==2024.1
urllib3==1.26.5
webdriver-manager==4.0.1
websocket-client==1.7.0
wsproto==1.2.0

# jq==1.7.1
# kubectl==1.30.2
# openshift-cli==4.14.5
# wkhtmltopdf==0.12.6
# MacOS: brew install Caskroom/cask/wkhtmltopdf
# Debian/Ubuntu: apt-get install wkhtmltopdf
# Windows: choco install wkhtmltopdf
#give necessary permission to record the screen
